# Conflicts:
#	src/main/java/com/evatool/analysis/api/controller/AnalysisControllerImpl.java
#	src/main/java/com/evatool/variants/events/listener/VariantAnalysisCreatedEventListener.java

+++ src/main/java/com/evatool/requirements/events/listener/RequirementEventListener.java
@@ -8,17 +8,15 @@ import com.evatool.global.event.impact.ImpactDeletedEvent;
 import com.evatool.global.event.impact.ImpactUpdatedEvent;
 import com.evatool.requirements.entity.RequirementDimension;
 import com.evatool.requirements.entity.RequirementsImpact;
-import com.evatool.requirements.entity.RequirementsVariant;
-import com.evatool.requirements.events.RequirementEventPublisher;
+import com.evatool.requirements.error.exceptions.EventEntityAlreadyExistsException;
+import com.evatool.requirements.error.exceptions.EventEntityDoesNotExistException;
+import com.evatool.requirements.repository.RequirementAnalysisRepository;
 import com.evatool.requirements.repository.RequirementDimensionRepository;
 import com.evatool.requirements.repository.RequirementsImpactsRepository;
 import com.evatool.requirements.repository.RequirementsVariantsRepository;
-import org.json.JSONException;
-import org.json.JSONObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.ApplicationEvent;
 import org.springframework.context.event.EventListener;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
@@ -26,7 +24,7 @@ import org.springframework.stereotype.Component;
 @Component
 public class RequirementEventListener {
 
-    Logger logger = LoggerFactory.getLogger(RequirementEventListener.class);
+    final Logger logger = LoggerFactory.getLogger(RequirementEventListener.class);
 
     @Autowired
     RequirementsImpactsRepository requirementsImpactsRepository;
@@ -34,14 +32,18 @@ public class RequirementEventListener {
     RequirementsVariantsRepository requirementsVariantsRepository;
     @Autowired
     RequirementDimensionRepository requirementDimensionRepository;
-
+    @Autowired
+    RequirementAnalysisRepository requirementAnalysisRepository;
 
     @EventListener
     @Async
     public void impactCreated(ImpactCreatedEvent event) {
         logger.info("impact created event ");
-        logger.debug("Event " + event.getSource() + " With Payload: " + event.getSource().toString() );
+        logger.debug("Event " + event.getClass() + " With Payload: " + event.getSource().toString() );
 
+        if (requirementsImpactsRepository.existsById(RequirementsImpact.fromJson(event.getJsonPayload()).getId())) {
+            throw new EventEntityAlreadyExistsException();
+        }
         requirementsImpactsRepository.save(RequirementsImpact.fromJson(event.getJsonPayload()));
     }
 
@@ -49,7 +51,10 @@ public class RequirementEventListener {
     @Async
     public void impactUpdated(ImpactUpdatedEvent event) {
         logger.info("Impact updated event");
-        logger.debug("Event " + event.getSource() + " With Payload: " + event.getSource().toString() );
+        logger.debug("Event " + event.getClass() + " With Payload: " + event.getSource().toString() );
+        if (!requirementsImpactsRepository.existsById(RequirementsImpact.fromJson(event.getJsonPayload()).getId())) {
+            throw new EventEntityDoesNotExistException();
+        }
         requirementsImpactsRepository.save(RequirementsImpact.fromJson(event.getJsonPayload()));
     }
 
@@ -57,14 +62,20 @@ public class RequirementEventListener {
     @Async
     public void impactDeleted(ImpactDeletedEvent event) {
         logger.info("Impact deleted event");
-        logger.debug("Event " + event.getSource() + " With Payload: " + event.getSource().toString() );
+        logger.debug("Event " + event.getClass() + " With Payload: " + event.getSource().toString() );
+        if (!requirementsImpactsRepository.existsById(RequirementsImpact.fromJson(event.getJsonPayload()).getId())) {
+            throw new EventEntityDoesNotExistException();
+        }
         requirementsImpactsRepository.delete(RequirementsImpact.fromJson(event.getJsonPayload()));
     }
     @EventListener
     @Async
     public void dimensionCreated(DimensionCreatedEvent event) {
         logger.info("dimension created event");
-        logger.debug("Event " + event.getSource() + " With Payload: " + event.getSource().toString() );
+        logger.debug("Event " + event.getClass() + " With Payload: " + event.getSource().toString() );
+        if (requirementDimensionRepository.existsById(RequirementDimension.fromJson(event.getJsonPayload()).getId())) {
+            throw new EventEntityAlreadyExistsException();
+        }
         requirementDimensionRepository.save(RequirementDimension.fromJson(event.getJsonPayload()));
     }
 
@@ -72,7 +83,10 @@ public class RequirementEventListener {
     @Async
     public void dimensionUpdated(DimensionUpdatedEvent event) {
         logger.info("dimension updated event");
-        logger.debug("Event " + event.getSource() + " With Payload: " + event.getSource().toString() );
+        logger.debug("Event " + event.getClass() + " With Payload: " + event.getSource().toString() );
+        if (!requirementDimensionRepository.existsById(RequirementDimension.fromJson(event.getJsonPayload()).getId())) {
+            throw new EventEntityDoesNotExistException();
+        }
         requirementDimensionRepository.save(RequirementDimension.fromJson(event.getJsonPayload()));
     }
 
@@ -80,12 +94,17 @@ public class RequirementEventListener {
     @Async
     public void dimensionDeleted(DimensionDeletedEvent event) {
         logger.info("dimension deleted event");
-        logger.debug("Event " + event.getSource() + " With Payload: " + event.getSource().toString() );
+        logger.debug("Event " + event.getClass() + " With Payload: " + event.getSource().toString() );
+        if (!requirementDimensionRepository.existsById(RequirementDimension.fromJson(event.getJsonPayload()).getId())) {
+            throw new EventEntityDoesNotExistException();
+        }
         requirementDimensionRepository.delete(RequirementDimension.fromJson(event.getJsonPayload()));
     }
+
+
     //events do not exist at the moment
 
-/**
+/*
  @EventListener
  @Async
     public void variantsCreated(ApplicationEvent event){
@@ -107,20 +126,28 @@ public class RequirementEventListener {
     logger.debug("Event " + event.getSource() + " With Payload: " + event.getSource().toString() );
     requirementsVariantsRepository.delete(RequirementsVariant.fromJson(event.getJsonPayload()));
     }
-
+*/
+    /*
     @EventListener
     @Async
-    public void analyseCreated(ApplicationEvent event){
+    public void analyseCreated(AnalysisCreatedEvent event){
         logger.info("analyse created event");
-        logger.debug("Event " + event.getSource() + " With Payload: " + event.getSource() );
-    }
+        logger.debug("Event " + event.getClass() + " With Payload: " + event.getMessage() );
+        if (requirementDimensionRepository.existsById(RequirementDimension.fromJson(event.getJsonPayload()).getId())) {
+            throw new EventEntityAlreadyExistsException();
+        }
+        requirementAnalysisRepository.save(RequirementsAnalysis.fromJson(event.getMessage()));
 
+    }
+    */
+/*
     @EventListener
     @Async
-    public void analyseDeleted(ApplicationEvent event){
+    public void analyseDeleted(AnalysisDeletedEvent analysisDeletedEvent){
         logger.info("analyse deleted event");
-        logger.debug("Event " + event.getSource() + " With Payload: " + event.getSource() );
+        logger.debug("Event " + analysisDeletedEvent.getClass() + " With Payload: " + analysisDeletedEvent.getMessage() );
+
 
     }
-**/
+*/
 }
